+++ lib/views/reactive_state.dart
@@ -4,6 +4,8 @@ import 'package:reactiv/dependency_management/dependency.dart';
 
 import 'bind_controller.dart';
 
+export 'bind_controller.dart';
+
 abstract class ReactiveState<T extends StatefulWidget, S extends ReactiveController> extends State<T> {
   final String? tag;